Clean some of the Hcal calibration code in the test area of Calibration/HcalCalibalgos
