+++ Calibration/HcalCalibAlgos/test/HcalHBHEMuonSimAnalyzer.cc
@@ -58,23 +58,21 @@ private:
   unsigned int matchId(const HcalDetId&, const HcalDetId&);
   double activeLength(const DetId&);
 
-  std::string g4Label_, ebLabel_, eeLabel_;
-  std::string hcLabel_;
-  int verbosity_, maxDepth_;
-  double etaMax_;
+  const std::string g4Label_, ebLabel_, eeLabel_, hcLabel_;
+  const int verbosity_, maxDepth_;
+  const double etaMax_, tMinE_, tMaxE_, tMinH_, tMaxH_;
+  const edm::EDGetTokenT<edm::SimTrackContainer> tok_SimTk_;
+  const edm::EDGetTokenT<edm::SimVertexContainer> tok_SimVtx_;
+  const edm::EDGetTokenT<edm::PCaloHitContainer> tok_caloEB_, tok_caloEE_;
+  const edm::EDGetTokenT<edm::PCaloHitContainer> tok_caloHH_;
+
+  const edm::ESGetToken<HcalDDDRecConstants, HcalRecNumberingRecord> tok_ddrec_;
+  const edm::ESGetToken<CaloGeometry, CaloGeometryRecord> tok_geom_;
+  const edm::ESGetToken<CaloTopology, CaloTopologyRecord> tok_caloTopology_;
+  const edm::ESGetToken<HcalTopology, HcalRecNumberingRecord> tok_topo_;
+  const edm::ESGetToken<MagneticField, IdealMagneticFieldRecord> tok_magField_;
+
   std::vector<HcalDDDRecConstants::HcalActiveLength> actHB_, actHE_;
-
-  double tMinE_, tMaxE_, tMinH_, tMaxH_;
-  edm::EDGetTokenT<edm::SimTrackContainer> tok_SimTk_;
-  edm::EDGetTokenT<edm::SimVertexContainer> tok_SimVtx_;
-  edm::EDGetTokenT<edm::PCaloHitContainer> tok_caloEB_, tok_caloEE_;
-  edm::EDGetTokenT<edm::PCaloHitContainer> tok_caloHH_;
-
-  edm::ESGetToken<HcalDDDRecConstants, HcalRecNumberingRecord> tok_ddrec_;
-  edm::ESGetToken<CaloGeometry, CaloGeometryRecord> tok_geom_;
-  edm::ESGetToken<CaloTopology, CaloTopologyRecord> tok_caloTopology_;
-  edm::ESGetToken<HcalTopology, HcalRecNumberingRecord> tok_topo_;
-  edm::ESGetToken<MagneticField, IdealMagneticFieldRecord> tok_magField_;
 
   const HcalDDDRecConstants* hcons_;
 
@@ -93,41 +91,35 @@ private:
   double hcalActiveLength_, hcalActiveLengthHot_;
 };
 
-HcalHBHEMuonSimAnalyzer::HcalHBHEMuonSimAnalyzer(const edm::ParameterSet& iConfig) {
-  usesResource(TFileService::kSharedResource);
+HcalHBHEMuonSimAnalyzer::HcalHBHEMuonSimAnalyzer(const edm::ParameterSet& iConfig) :
+  g4Label_(iConfig.getParameter<std::string>("ModuleLabel")),
+  ebLabel_(iConfig.getParameter<std::string>("EBCollection")),
+  eeLabel_(iConfig.getParameter<std::string>("EECollection")),
+  hcLabel_(iConfig.getParameter<std::string>("HCCollection")),
+  verbosity_(iConfig.getUntrackedParameter<int>("Verbosity", 0)),
+  maxDepth_(iConfig.getUntrackedParameter<int>("MaxDepth", 4)),
+  etaMax_(iConfig.getUntrackedParameter<double>("EtaMax", 3.0)),
+  tMinE_(iConfig.getUntrackedParameter<double>("TimeMinCutECAL", -500.)),
+  tMaxE_(iConfig.getUntrackedParameter<double>("TimeMaxCutECAL", 500.)),
+  tMinH_(iConfig.getUntrackedParameter<double>("TimeMinCutHCAL", -500.)),
+  tMaxH_(iConfig.getUntrackedParameter<double>("TimeMaxCutHCAL", 500.)),
+  tok_SimTk_(consumes<edm::SimTrackContainer>(edm::InputTag(g4Label_))),
+  tok_SimVtx_(consumes<edm::SimVertexContainer>(edm::InputTag(g4Label_))),
+  tok_caloEB_(consumes<edm::PCaloHitContainer>(edm::InputTag(g4Label_, ebLabel_))),
+  tok_caloEE_(consumes<edm::PCaloHitContainer>(edm::InputTag(g4Label_, eeLabel_))),
+  tok_caloHH_(consumes<edm::PCaloHitContainer>(edm::InputTag(g4Label_, hcLabel_))),
+  tok_ddrec_(esConsumes<HcalDDDRecConstants, HcalRecNumberingRecord, edm::Transition::BeginRun>()),
+  tok_geom_(esConsumes<CaloGeometry, CaloGeometryRecord>()),
+  tok_caloTopology_(esConsumes<CaloTopology, CaloTopologyRecord>()),
+  tok_topo_(esConsumes<HcalTopology, HcalRecNumberingRecord>()),
+  tok_magField_(esConsumes<MagneticField, IdealMagneticFieldRecord>()) {
 
   //now do what ever initialization is needed
-  g4Label_ = iConfig.getParameter<std::string>("ModuleLabel");
-  ebLabel_ = iConfig.getParameter<std::string>("EBCollection");
-  eeLabel_ = iConfig.getParameter<std::string>("EECollection");
-  hcLabel_ = iConfig.getParameter<std::string>("HCCollection");
-  verbosity_ = iConfig.getUntrackedParameter<int>("Verbosity", 0);
-  maxDepth_ = iConfig.getUntrackedParameter<int>("MaxDepth", 4);
-  etaMax_ = iConfig.getUntrackedParameter<double>("EtaMax", 3.0);
-  tMinE_ = iConfig.getUntrackedParameter<double>("TimeMinCutECAL", -500.);
-  tMaxE_ = iConfig.getUntrackedParameter<double>("TimeMaxCutECAL", 500.);
-  tMinH_ = iConfig.getUntrackedParameter<double>("TimeMinCutHCAL", -500.);
-  tMaxH_ = iConfig.getUntrackedParameter<double>("TimeMaxCutHCAL", 500.);
-
-  tok_SimTk_ = consumes<edm::SimTrackContainer>(edm::InputTag(g4Label_));
-  tok_SimVtx_ = consumes<edm::SimVertexContainer>(edm::InputTag(g4Label_));
-  tok_caloEB_ = consumes<edm::PCaloHitContainer>(edm::InputTag(g4Label_, ebLabel_));
-  tok_caloEE_ = consumes<edm::PCaloHitContainer>(edm::InputTag(g4Label_, eeLabel_));
-  tok_caloHH_ = consumes<edm::PCaloHitContainer>(edm::InputTag(g4Label_, hcLabel_));
-  if (maxDepth_ > depthMax_)
-    maxDepth_ = depthMax_;
-  else if (maxDepth_ < 1)
-    maxDepth_ = 4;
+  usesResource(TFileService::kSharedResource);
 
   edm::LogVerbatim("HBHEMuon") << "Labels: " << g4Label_ << ":" << ebLabel_ << ":" << eeLabel_ << ":" << hcLabel_
                                << "\nVerbosity " << verbosity_ << " MaxDepth " << maxDepth_ << " Maximum Eta "
                                << etaMax_ << " tMin|tMax " << tMinE_ << ":" << tMaxE_ << ":" << tMinH_ << ":" << tMaxH_;
-
-  tok_ddrec_ = esConsumes<HcalDDDRecConstants, HcalRecNumberingRecord, edm::Transition::BeginRun>();
-  tok_geom_ = esConsumes<CaloGeometry, CaloGeometryRecord>();
-  tok_caloTopology_ = esConsumes<CaloTopology, CaloTopologyRecord>();
-  tok_topo_ = esConsumes<HcalTopology, HcalRecNumberingRecord>();
-  tok_magField_ = esConsumes<MagneticField, IdealMagneticFieldRecord>();
 }
 
 void HcalHBHEMuonSimAnalyzer::analyze(const edm::Event& iEvent, const edm::EventSetup& iSetup) {
